refactoring tests - no new features
 - correcting tests that were incorrect or insufficient
 - uncommenting and fixing numerous test cases
 - extract reused code into common js module
 - rename test source files to specify the plugin name where appropriate
 - add more debugging into some plugins
 - make the unix formatter slightly less taciturn
