+++ lib/package/bundleLinter.js
@@ -178,20 +178,24 @@ var getFormatter = function(format) {
 const bfnName = (term) =>
   (term == 'Bundle') ? 'onBundle' : pluralize('on' + term, 2);
 
-const pluginIdRe = new RegExp('^[A-Z]{2}[0-9]{3}$');
+const pluginIdRe1 = new RegExp('^[A-Z]{2}[0-9]{3}$');
+const pluginIdRe2 = new RegExp('^_.+\.js$');
 
 /* exposed for testing */
 const resolvePlugin = idOrFilename => {
         if (idOrFilename.endsWith(".js")) {
           if (idOrFilename.indexOf('/') < 0) {
-            debug(`executePlugin file(${idOrFilename}) , prepending path...`);
-            return path.resolve(getPluginPath(), idOrFilename);
+            if ( ! pluginIdRe2.test(idOrFilename)) {
+              debug(`executePlugin file(${idOrFilename}) , prepending path...`);
+              return path.resolve(getPluginPath(), idOrFilename);
+            }
+            else return null;
           }
           else {
             return idOrFilename;
           }
         }
-        else if (pluginIdRe.test(idOrFilename)) {
+        else if (pluginIdRe1.test(idOrFilename)) {
           let p = fs.readdirSync(path.resolve(getPluginPath())).find( p => p.startsWith(idOrFilename));
           return p ? path.resolve(getPluginPath(), p) : null;
         }
