+++ lib/package/plugins/PO001-threatProtectionJSONConditionCheck.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2019-2020 Google LLC
+  Copyright 2019-2021 Google LLC
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
   limitations under the License.
 */
 
-//threatProtectionJSONConditionCheck
 //| &nbsp; |:white_medium_square:| PO001 | JSON Threat Protection |  A check for a body element must be performed before policy execution. |
-//| &nbsp; |:white_medium_square:| PO002 | XML Threat Protection |  A check for a body element must be performed before policy execution. |
+
 
 const ruleId = require("../myUtil.js").getRuleId(),
       debug = require("debug")("apigeelint:" + ruleId);
@@ -32,45 +31,11 @@ const plugin = {
     enabled: true
   },
   condRegExp =
-    "(response.content|response.form|request.content|request.form|message.content|message.form|message.verb|request.verb)";
+    "(response.content|response.form|request.content|request.form|message.content|message.form|request.verb|message.verb)";
 
-const onPolicy = function(policy, cb) {
-  var hadWarning = false;
-  if (policy.getType() === "JSONThreatProtection") {
-    if (policy.getSteps().length > 0) {
-      var missingBodyCheck = false,
-        steps = policy.getSteps();
-      //get steps
-      //check each step for a condition on body
-      steps.forEach(function(step) {
-        let condition = step.getCondition();
-        if (!condition || !condition.getExpression().match(condRegExp)) {
-          missingBodyCheck = true;
+const checker = require('./_policyConditionCheck.js');
 
-          //is the parent a flow we might revert the decision if it has an appropriate condition
-          if (step.parent && step.parent.getType() === "Flow") {
-            let condition = step.parent.getCondition();
-            if (condition && condition.getExpression().match(condRegExp)) {
-              missingBodyCheck = false;
-            }
-          }
-        }
-      });
-
-      if (missingBodyCheck) {
-        hadWarning = true;
-        policy.addMessage({
-          plugin,
-          message:
-            "An appropriate check for a message body was not found on the enclosing Step or Flow."
-        });
-      }
-    }
-  }
-  if (typeof(cb) == 'function') {
-    cb(null, hadWarning);
-  }
-};
+const onPolicy = checker(plugin, "JSONThreatProtection", condRegExp, debug);
 
 module.exports = {
   plugin,
