+++ lib/package/plugins/PO002-threatProtectionXMLConditionCheck.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2019-2020 Google LLC
+  Copyright 2019-2021 Google LLC
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
   limitations under the License.
 */
 
-//threatProtectionXMLConditionCheck
+
 //| &nbsp; |:white_medium_square:| PO002 | XML Threat Protection |  A check for a body element must be performed before policy execution. |
 
 const ruleId = require("../myUtil.js").getRuleId(),
@@ -33,43 +33,9 @@ const plugin = {
   condRegExp =
     "(response.content|response.form|request.content|request.form|message.content|message.form|message.verb|request.verb)";
 
-var onPolicy = function(policy, cb) {
-      let hadWarning = false;
-  if (policy.getType() === "XMLThreatProtection") {
-    if (policy.getSteps().length > 0) {
-      var missingBodyCheck = false,
-        steps = policy.getSteps();
-      //get steps
-      //check each step for a condition on body
-      steps.forEach(function(step) {
-        let condition = step.getCondition();
-        if (!condition || !condition.getExpression().match(condRegExp)) {
-          missingBodyCheck = true;
+const checker = require('./_policyConditionCheck.js');
 
-          //is the parent a flow we might revert the decision if it has an appropriate condition
-          if (step.parent && step.parent.getType() === "Flow") {
-            let condition = step.parent.getCondition();
-            if (condition && condition.getExpression().match(condRegExp)) {
-              missingBodyCheck = false;
-            }
-          }
-        }
-      });
-
-      if (missingBodyCheck) {
-        hadWarning = true;
-        policy.addMessage({
-          plugin,
-          message:
-            "An appropriate check for a message body was not found on the enclosing Step or Flow."
-        });
-      }
-    }
-  }
-  if (typeof(cb) == 'function') {
-    cb(null, hadWarning);
-  }
-};
+const onPolicy = checker(plugin, "XMLThreatProtection", condRegExp, debug);
 
 module.exports = {
   plugin,
