+++ lib/package/plugins/PO003-extractJSONConditionCheck.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2019-2020 Google LLC
+  Copyright 2019-2021 Google LLC
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -33,59 +33,85 @@ const plugin = {
       };
 
 const onPolicy = function(policy, cb) {
-  let condRegExp =
-    "(response.content|response.form|request.content|request.form|message.content|message.form|message.verb|request.verb)";
+        let condRegExp =
+        "(response\.content|response\.form|request\.content|request\.form|message\.content|message\.form|message\.verb|request\.verb)";
 
-  let hadWarning = false;
+  let flagged = false;
   if (policy.getType() === "ExtractVariables") {
-    let jsonPayload = xpath.select(
-      "/ExtractVariables/JSONPayload/text()",
-      policy.getElement()
-    );
+    debug('found ExtractVariables policy');
 
-    let sourceElement = xpath.select(
-        "/ExtractVariables/Source/text()",
-        policy.getElement()
-    );
+    if (policy.getSteps().length > 0) {
+      debug(`policy has ${policy.getSteps().length} steps`);
+      let jsonPayload = xpath.select(
+            "/ExtractVariables/JSONPayload/text()",
+            policy.getElement()
+          );
 
-    if (sourceElement.length) {
-      let source = sourceElement[0].data;
-      condRegExp = !source.match(condRegExp) ? source : condRegExp;
-    }
+      if (jsonPayload.length > 0) {
+        debug('is JSONPayload');
 
-    if (jsonPayload.length > 0 && policy.getSteps().length > 0) {
-      let hasBodyChecks =
-        policy.getSteps().every(step => {
-          let condition = step.getCondition();
-          if (condition && condition.getExpression().match(condRegExp)) {
-            return true;
-          }
+        let sourceElement = xpath.select(
+              "/ExtractVariables/Source/text()",
+              policy.getElement()
+            );
 
-          let parent = step.parent;
-          if (parent && parent.getType() === "FlowPhase") {
-            parent = parent.parent;
-          }
-          if (parent && parent.getType() === "Flow") {
-            condition = step.parent.getCondition();
+        if (sourceElement.length) {
+          let source = sourceElement[0].data;
+          condRegExp = !source.match(condRegExp) ? `${source}\.content` : condRegExp;
+          debug(`change condRegExp: ${condRegExp}`);
+        }
+
+        let hasBodyChecks =
+          policy.getSteps().every(step => {
+            // let util = require('util');
+            // debug('step ' + util.format(step));
+            let condition = step.getCondition();
+            if (condition) {
+              debug(`cond expression: [${condition.getExpression()}]`);
+            }
+            else {
+              debug('no Condition');
+            }
             if (condition && condition.getExpression().match(condRegExp)) {
+              debug('sufficient condition');
               return true;
             }
-          }
-          return false;
-        });
 
-      if ( ! hasBodyChecks) {
-        hadWarning = true;
-        policy.addMessage({
-          plugin,
-          message:
-          "An appropriate check for a message body was not found on the enclosing Step or Flow."
-        });
+            let parent = step.parent;
+            if (parent && parent.getType() === "FlowPhase") {
+              parent = parent.parent;
+            }
+            if (parent && parent.getType() === "Flow") {
+              debug('parent is Flow');
+              condition = step.parent.getCondition();
+              if (condition && condition.getExpression().match(condRegExp)) {
+                debug('parent has sufficient condition');
+                return true;
+              }
+            }
+            debug('missing or insufficient condition');
+            return false;
+          });
+
+        if ( ! hasBodyChecks) {
+          flagged = true;
+          policy.addMessage({
+            plugin,
+            message:
+            "An appropriate check for a message body was not found on the enclosing Step or Flow."
+          });
+        }
       }
+      else {
+        debug('not a JSONPayload')
+      }
+    }
+    else {
+      debug('not attached')
     }
   }
   if (typeof(cb) == 'function') {
-    cb(null, hadWarning);
+    cb(null, flagged);
   }
 };
