+++ test/specs/BN001-xsdResourceTest.js
