+++ test/specs/PO026-assignVariableHygieneInRaiseFault.js
