+++ test/specs/TD002-edgemicroNot.js
@@ -1,8 +1,20 @@
-// assignToHygiene.js
-// ------------------------------------------------------------------
+/*
+  Copyright 2019-2021 Google LLC
 
-/* jshint esversion:9, node:true, strict:implied */
-/* global process, console, Buffer, describe, it */
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+      https://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+*/
+
+/* global describe, it */
 
 const assert = require("assert"),
       path = require("path"),
