+++ test/specs/testTruthTableVarComps.js
@@ -1,5 +1,5 @@
 /*
-  Copyright 2019-2020 Google LLC
+  Copyright 2019-2021 Google LLC
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -13,6 +13,8 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
+
+/* global describe, it */
 
 const assert = require("assert"),
       TruthTable = require("../../lib/package/TruthTable.js"),
